Major rework of the SDFITS writer.  The version is more like the FITS IDI writer and create output that can be read into CASA.  Modified the TBN correlators to correlate for less than 1 second.  Fixed the imaging script to use the inner 2/3 of the bandpass.

+++ scripts/IOC/correlateTBN.py
@@ -102,7 +102,7 @@ def parseConfig(args):
 		elif opt in ('-l', '--fft-length'):
 			config['LFFT'] = int(value)
 		elif opt in ('-t', '--avg-time'):
-			config['avgTime'] = int(value)
+			config['avgTime'] = float(value)
 		elif opt in ('-s', '--samples'):
 			config['samples'] = int(value)
 		elif opt in ('-o', '--offset'):
